Create BFS
The following are the steps involved in employing breadth-first search to explore a graph:

Take the data for the graph's adjacency matrix or adjacency list.
Create a queue and fill it with items.
Activate the root node (meaning that get the root node at the beginning of the queue).
Dequeue the queue's head (or initial element), then enqueue all of the queue's nearby nodes from left to right. Simply dequeue the head and resume the operation if a node has no nearby nodes that need to be investigated. (Note: If a neighbor emerges that has previously been investigated or is in the queue, don't enqueue it; instead, skip it.)
Continue in this manner until the queue is empty.